refactor: styled-component의 파일 위치 이동
#11
